\name{notdocumentedyet}
\alias{notdocumentedyet}

\alias{createBlob}
\alias{deleteBlob}
\alias{getBlob}
\alias{loadBlob}
\alias{getmd5}
\alias{checkmd5}
\alias{read.ini}
\alias{write.ini}
\alias{storeIniObj}
\alias{getMetaObjName}
\alias{getMetaObj}

%
%
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Undocumented and Internally Used Functions and Classes }
\description{
Currently undocumented functions

}
%\usage{
%normal1(lmean="identity", lsd="loge", zero=NULL)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{lmean}{
%  Link function applied to the mean.
%  See \code{\link{Links}} for more choices.
%
%  }
%}
\details{
  In the \pkg{pgblobs} package there are currently many
  objects/methods/classes which are currently internal and/or
  undocumented.

}
\value{
  Each objects/methods/classes may or may not have its own individual value.
  These will be documented over time.

}
%\references{
%}

\author{ Vectrun }
%\note{
%
%}
%\seealso{
%    \code{gaussianff},
%    \code{\link{posnormal1}}.
%}
%\examples{
%}
\keyword{undocumented}







