\name{pgoblobs-package}
\alias{pgblobs-package}
\alias{pgblobs}
\docType{package}
\title{Postgresql - R objects blobs
}
\description{
    Package to maintain blobs in pgobjects postgresql database
}
\details{
\tabular{ll}{
Package: \tab pgblobs\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2016-02-25\cr
License: \tab  GPL v3.0\cr
LazyLoad: \tab yes\cr
}
With this package it is possible to store blobs with pgobjects.
}
\author{
Vectrun

Maintainer: Vectrun <vectrun@gmail.com>
}
\references{
    none
}
\keyword{ package }
\seealso{
}
\examples{
}
